Add retrieveById method and update socket.on 'message' event
A retrieveById method was added to fetch specific messages by their ID from the database. Moreover, the socket.on 'message' event functionality has been expanded. Now, whenever a new message is sent, the user's actual info is fetched from the DB, the new message is saved to the DB and the result is loaded, and this message is then emitted to all participants in the chat.